divide sub controllers.
